Replace all buttons with new version
