+++ shared/@hedvig-ui/Button/button.stories.tsx
@@ -1,4 +1,4 @@
-import { Button, buttonColorMap, ButtonProps, lightTheme } from '@hedvig-ui'
+import { Button, ButtonProps } from '@hedvig-ui'
 import { action } from '@storybook/addon-actions'
 import { boolean, select } from '@storybook/addon-knobs'
 import React from 'react'
@@ -9,9 +9,12 @@ export default {
   component: Button,
 }
 
-const colors: ReadonlyArray<NonNullable<
-  ButtonProps['variation']
->> = Object.keys(buttonColorMap(lightTheme)) as any
+const variants: ReadonlyArray<NonNullable<ButtonProps['variant']>> = [
+  'primary',
+  'secondary',
+  'tertiary',
+]
+
 const sizes: ReadonlyArray<NonNullable<ButtonProps['size']>> = [
   'small',
   'medium',
@@ -22,10 +25,9 @@ export const Text = () => {
   return (
     <Button
       onClick={action('clicked')}
-      variation={select('variation', colors, 'primary')}
+      variant={select('variant', variants, 'primary')}
       size={select('Size', sizes, 'medium')}
-      basic={boolean('Basic', false)}
-      loading={boolean('Loading', false)}
+      disabled={boolean('Disabled', false)}
     >
       Hello Button
     </Button>
@@ -35,9 +37,8 @@ export const Text = () => {
 export const Emoji: React.FC & WithStory = () => (
   <Button
     onClick={action('clicked')}
-    variation={select('variation', colors, 'primary')}
+    variant={select('variation', variants, 'primary')}
     size={select('Size', sizes, 'medium')}
-    basic={boolean('Basic', false)}
   >
     <span role="img" aria-label="so cool">
       😀 😎 👍 💯
