+++ src/features/claims/claim-details/components/ClaimReserveForm.tsx
@@ -20,7 +20,6 @@ const ClaimReserveForm: React.FC<{ claimId: string }> = ({ claimId }) => {
       />
       <Spacing top="small" />
       <Button
-        variation="primary"
         disabled={!isStringNumber(value) || value === '' || loading}
         onClick={async () => {
           await updateReserve({
