+++ src/features/claims/claim-details/components/ClaimType.tsx
@@ -174,9 +174,7 @@ const ClaimTypeDataForm: React.FC<{ type: any; claimId: string }> = ({
       )}
       <Spacing top="medium" />
       <Button
-        fullWidth
         disabled={formData === savedFormData || loading}
-        variation="primary"
         onClick={() => {
           setClaimInformation({
             variables: {
