+++ src/features/member/tabs/claims-tab/ClaimsTab.tsx
@@ -65,8 +65,7 @@ export const ClaimsTab: React.FC<{
             />
             <Button
               disabled={createClaimLoading}
-              variation="danger"
-              color="danger"
+              variant="secondary"
               onClick={() => {
                 setShowForm(false)
                 setClaimDate(new Date())
@@ -79,8 +78,6 @@ export const ClaimsTab: React.FC<{
               disabled={
                 claimSource === null || claimDate === null || createClaimLoading
               }
-              variation="success"
-              color="success"
               onClick={async () => {
                 if (claimSource === null || claimDate === null) {
                   return
@@ -113,18 +110,13 @@ export const ClaimsTab: React.FC<{
                 )
               }}
             >
-              Add
+              Create
             </Button>
           </FormWrapper>
         ) : (
-          <Button
-            variation="primary"
-            color="primary"
-            onClick={() => setShowForm(true)}
-            style={{ height: 39 }}
-          >
-            Add new claim
-          </Button>
+          <div>
+            <Button onClick={() => setShowForm(true)}>Create new claim</Button>
+          </div>
         )}
       </HeaderWrapper>
       <MemberClaimsList memberId={memberId} />
