+++ src/features/member/tabs/contracts-tab/contract/master-inception.tsx
@@ -37,11 +37,7 @@ export const MasterInception: React.FC<{
   return (
     <>
       {!datePickerEnabled && !contract.terminationDate && (
-        <Button
-          halfWidth
-          variation="success"
-          onClick={() => setDatePickerEnabled(true)}
-        >
+        <Button status="success" onClick={() => setDatePickerEnabled(true)}>
           Activate
         </Button>
       )}
@@ -53,7 +49,6 @@ export const MasterInception: React.FC<{
           <DateTimePicker date={activeFrom} setDate={setActiveFrom} />
           <ButtonsGroup>
             <Button
-              fullWidth
               disabled={activateContractLoading}
               onClick={() => {
                 const confirmMessage = `Are you sure you want to activate this contract with master inception of ${format(
@@ -77,11 +72,10 @@ export const MasterInception: React.FC<{
                   )
                 })
               }}
-              variation="success"
             >
               Confirm
             </Button>
-            <Button fullWidth onClick={() => reset()}>
+            <Button variant="tertiary" onClick={() => reset()}>
               Cancel
             </Button>
           </ButtonsGroup>
