+++ src/features/member/tabs/contracts-tab/contract/termination-date.tsx
@@ -66,15 +66,13 @@ export const TerminationDate: React.FC<{
             </Spacing>
             <ButtonsGroup>
               <Button
-                fullWidth
-                variation="secondary"
+                variant="secondary"
                 onClick={() => setDatePickerEnabled(true)}
               >
                 Change
               </Button>
               <Button
-                fullWidth
-                variation="success"
+                status="success"
                 disabled={revertTerminationLoading}
                 onClick={() => {
                   confirm('Are you want to revert the termination?').then(
@@ -120,8 +118,6 @@ export const TerminationDate: React.FC<{
             </Spacing>
             <ButtonsGroup>
               <Button
-                fullWidth
-                variation="primary"
                 disabled={changeTerminationDateLoading}
                 onClick={() => {
                   const confirmMessage = `Are you sure you want to change the termination date from ${
@@ -157,7 +153,7 @@ export const TerminationDate: React.FC<{
               >
                 Confirm
               </Button>
-              <Button fullWidth onClick={() => reset()}>
+              <Button variant="tertiary" onClick={() => reset()}>
                 Cancel
               </Button>
             </ButtonsGroup>
@@ -169,11 +165,7 @@ export const TerminationDate: React.FC<{
   return (
     <>
       {!datePickerEnabled && (
-        <Button
-          halfWidth
-          variation="danger"
-          onClick={() => setDatePickerEnabled(true)}
-        >
+        <Button status="danger" onClick={() => setDatePickerEnabled(true)}>
           Terminate contract
         </Button>
       )}
@@ -194,8 +186,7 @@ export const TerminationDate: React.FC<{
           </Spacing>
           <ButtonsGroup>
             <Button
-              fullWidth
-              variation="danger"
+              status="danger"
               disabled={terminationReason === null || terminateContractLoading}
               onClick={() => {
                 const confirmedMsg = `Are you sure you want to terminate this contract with the termination date ${format(
@@ -240,7 +231,7 @@ export const TerminationDate: React.FC<{
             >
               Terminate
             </Button>
-            <Button fullWidth onClick={() => reset()}>
+            <Button variant="tertiary" onClick={() => reset()}>
               Cancel
             </Button>
           </ButtonsGroup>
