+++ src/features/member/tabs/quote-tab/QuotePrice.tsx
@@ -113,10 +113,10 @@ export const QuotePrice = ({ quote }: Props) => {
                 onChange={onPriceChange}
               />
             </PriceInput>
-            <Button variation="icon" type="submit">
+            <Button type="submit">
               <SubmitButton size="1.6em" />
             </Button>
-            <Button variation="icon" onClick={onCancel}>
+            <Button onClick={onCancel}>
               <CancelButton size="1.6em" />
             </Button>
           </AlignCenter>
@@ -124,7 +124,7 @@ export const QuotePrice = ({ quote }: Props) => {
       ) : (
         <AlignCenter>
           <DisplayPrice>{formattedPrice}</DisplayPrice>
-          <Button variation="icon" onClick={() => setEditPrice(true)}>
+          <Button onClick={() => setEditPrice(true)}>
             <PencilFill />
           </Button>
         </AlignCenter>
