+++ src/features/questions/questions-list/QuestionGroupInfo.tsx
@@ -56,10 +56,9 @@ export const QuestionGroupInfo = ({ questionGroup }) => {
                 }
               >
                 <Button
-                  variation="icon"
+                  variant="tertiary"
                   disabled={member.claims.length > 1}
                   onClick={() => history.push(`/claims/${member.claims[0].id}`)}
-                  color="primary"
                   size="large"
                   style={{ padding: '0em', marginRight: '1em' }}
                 >
