+++ src/features/tools/campaign-codes/forms/MonthlyPercentageForm.tsx
@@ -175,8 +175,6 @@ export const MonthlyPercentageForm: React.FC = () => {
       <Spacing top="small" />
       <div>
         <Button
-          variation="primary"
-          loading={loading}
           disabled={loading || !formLooksGood(formData)}
           onClick={() => {
             confirm(`Create new campaign code "${formData.code}"?`).then(() => {
