+++ src/features/tools/employees/components/EmployeeTable.tsx
@@ -95,7 +95,7 @@ export const EmployeeTable: React.FC<{
 
               return (
                 <Table.Row key={id}>
-                  <Table.Cell width={5}>{email}</Table.Cell>
+                  <Table.Cell width={3}>{email}</Table.Cell>
                   <Table.Cell width={3}>
                     {scopes.includes('employees:manage') ? (
                       <Dropdown
@@ -116,10 +116,9 @@ export const EmployeeTable: React.FC<{
                   <Table.Cell width={3}>
                     {dateTimeFormatter(firstGrantedAt, 'yyyy-MM-dd HH:mm')}
                   </Table.Cell>
-                  <Table.Cell width={1}>
+                  <Table.Cell width={2}>
                     <ButtonsGroup>
                       <Button
-                        variation="primary"
                         disabled={
                           !selectedRoles[id] ||
                           selectedRoles[id] === currentRoles[id] ||
@@ -141,10 +140,10 @@ export const EmployeeTable: React.FC<{
                           )
                         }
                       >
-                        Update Role
+                        Update role
                       </Button>
                       <Button
-                        variation="danger"
+                        variant="tertiary"
                         disabled={
                           removeEmployeeLoading ||
                           !scopes.includes('employees:manage')
