+++ src/features/tools/norwegian-tariff-editor/postal-codes-editor.tsx
@@ -24,8 +24,7 @@ export const PostalCodesEditor: React.FC = () => {
       />
       <Spacing top>
         <Button
-          fullWidth
-          variation="secondary"
+          variant="secondary"
           disabled={loading}
           onClick={() => {
             confirm('Are you sure you want to add the postal codes?').then(
