+++ src/pages/DashboardPage.tsx
@@ -2,7 +2,6 @@ import styled from '@emotion/styled'
 import {
   Badge,
   Capitalized,
-  Card,
   CasualList,
   CasualListItem,
   FadeIn,
@@ -14,10 +13,8 @@ import {
 import { changelog } from 'changelog'
 import { differenceInCalendarDays, format } from 'date-fns'
 import { NumberMemberGroupsRadioButtons } from 'features/questions/number-member-groups-radio-buttons'
-import { Button } from 'features/tools/employees/button'
 import { useDashboardNumbers } from 'graphql/use-dashboard-numbers'
 import React from 'react'
-import { Shield, StarFill } from 'react-bootstrap-icons'
 import { Link } from 'react-router-dom'
 import { useGetMeQuery } from 'types/generated/graphql'
 import { useInsecurePersistentState } from 'utils/state'
@@ -74,35 +71,6 @@ export const DashboardPage: React.FC = () => {
 
   return (
     <Wrapper>
-      <div>
-        <Button icon={<Shield />} variant={'primary'} status={'success'}>
-          Hello primary
-        </Button>
-        <Button variant={'primary'} status={'success'}>
-          Hello primary
-        </Button>
-        <Button
-          icon={<StarFill />}
-          size="small"
-          variant={'secondary'}
-          status={'warning'}
-        >
-          Hello secondary
-        </Button>
-        <Button
-          icon={<StarFill />}
-          size="large"
-          variant={'tertiary'}
-          status={'danger'}
-        >
-          Hello tertiary
-        </Button>
-      </div>
-      <Card>
-        <Button variant={'primary'}>Hello primary</Button>
-        <Button variant={'secondary'}>Hello secondary</Button>
-        <Button variant={'tertiary'}>Hello tertiary</Button>
-      </Card>
       <Spacing bottom top>
         {data?.me && (
           <MainHeadline>
