+++ src/pages/tools/NorwegianTariffCreatorPage.tsx
@@ -73,8 +73,6 @@ export const NorwegianTariffCreatorPage: React.FC = () => {
       ))}
       <Spacing top>
         <Button
-          fullWidth
-          variation="primary"
           disabled={loading}
           onClick={() => {
             confirm('Are you sure you want to create the price engine?').then(
