+++ src/utils/hooks/modal-hook.tsx
@@ -48,15 +48,10 @@ export const ConfirmDialogComponent: React.FC<{
       <div>
         <h3>{content}</h3>
         <ConfirmButtons>
-          <Button
-            autoFocus={true}
-            fullWidth
-            variation="success"
-            onClick={confirm}
-          >
+          <Button status="success" onClick={confirm}>
             Confirm
           </Button>
-          <Button fullWidth variation="secondary" onClick={close}>
+          <Button variant="tertiary" onClick={close}>
             Cancel
           </Button>
         </ConfirmButtons>
